Fix/develop bugs

+++ src/services/clinic/clinicDelete.service.ts
@@ -16,15 +16,22 @@ const clinicDeleteService = async (id: string, userId: string) => {
   const clinicsDoctors = await clinicsDoctorsRepository.findOne({
     where: {
       clinic: { id },
+      doctor: { id: userId },
     },
     relations: {
       clinic: true,
+      doctor: true,
     },
   });
+
+  if (!clinicsDoctors) {
+    throw new AppError("Clinic has already been removed", 400);
+  }
 
   await clinicsDoctorsRepository.delete({
     id: clinicsDoctors!.id,
   });
+
   return "Clinic removed successfully";
 };
 export default clinicDeleteService;
